Move NetplaySetupScreen to its own file
